+++ forui_internal_gen/lib/forui_internal_gen.dart
@@ -1,0 +1,12 @@
+import 'package:build/build.dart';
+import 'package:dart_style/dart_style.dart';
+import 'package:source_gen/source_gen.dart';
+
+import 'package:forui_internal_gen/src/style_generator.dart';
+
+/// Builds generators for `build_runner` to run
+Builder styleBuilder(BuilderOptions options) => PartBuilder(
+    [StyleGenerator()],
+    '.style.dart',
+    formatOutput: (generated, version) => DartFormatter(pageWidth: 120, languageVersion: version).format(generated),
+  );